requirement.txt
Pillow==10.0
numpy==1.26.3
python-debouncer==0.1.3
names==0.3.0
munch
deepmerge
nicegui[redis]
